+++ src/android/app/src/main/java/org/citra/citra_emu/features/settings/utils/SettingsFile.java
@@ -35,8 +35,6 @@ public final class SettingsFile {
 
     public static final String KEY_CPU_JIT = "use_cpu_jit";
 
-    public static final String KEY_INACCURATE_EMULATION = "inaccurate_emulation";
-
     public static final String KEY_DESIGN = "design";
 
     public static final String KEY_USE_PRESENT_THREAD = "use_present_thread";
@@ -67,7 +65,7 @@ public final class SettingsFile {
     public static final String KEY_BACKGROUND_GREEN = "bg_green";
     public static final String KEY_RENDER_3D = "render_3d";
     public static final String KEY_FACTOR_3D = "factor_3d";
-    // public static final String KEY_PP_SHADER_NAME = "pp_shader_name";
+    public static final String KEY_PP_SHADER_NAME = "pp_shader_name";
     // public static final String KEY_FILTER_MODE = "filter_mode";
     public static final String KEY_TEXTURE_FILTER_NAME = "texture_filter_name";
     public static final String KEY_USE_ASYNCHRONOUS_GPU_EMULATION = "use_asynchronous_gpu_emulation";
